classifier/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
altgraph=0.17.2=pypi_0
astor=0.8.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
ca-certificates=2022.3.29=haa95532_0
certifi=2021.5.30=py36haa95532_0
colorama=0.4.4=pyhd3eb1b0_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=pypi_0
gast=0.2.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.44.0=pypi_0
h5py=2.10.0=pypi_0
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
importlib-metadata=4.8.3=pypi_0
intel-openmp=2022.0.0=haa95532_3663
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py36_0
jpeg=9d=h2bbff1b_0
keras=2.8.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py36hd77b12b_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.6=pypi_0
matplotlib=3.3.4=py36haa95532_0
matplotlib-base=3.3.4=py36h49ac443_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
numpy=1.18.5=pypi_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.46=py36_0
opencv=3.4.2=py36h40b0b35_0
openssl=1.1.1n=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pefile=2022.5.30=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py36h4fa10fc_0
pip=21.2.2=py36haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
py-opencv=3.4.2=py36hc319ecb_0
pygments=2.11.2=pyhd3eb1b0_0
pyinstaller=4.10=pypi_0
pyinstaller-hooks-contrib=2022.0=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
python=3.6.13=h3758d61_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pywin32-ctypes=0.2.0=pypi_0
qt=5.9.7=vc14h73c81de_0
setuptools=58.0.4=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h2bbff1b_0
tensorboard=1.15.0=pypi_0
tensorflow-directml=1.15.5=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h2bbff1b_0
tornado=6.1=py36h2bbff1b_0
traitlets=4.3.3=py36haa95532_0
typing-extensions=4.1.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.14.0=pypi_0
xz=5.2.5=h62dcd97_0
zipp=3.6.0=pypi_0
zlib=1.2.11=hbd8134f_5
zstd=1.4.9=h19a0ad4_0
